fixed some issues with login and modularized login routing
