+++ package.json
@@ -11,6 +11,7 @@
   },
   "license": "ISC",
   "dependencies": {
+    "bcrypt": "^3.0.5",
     "bcryptjs": "^2.4.3",
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
