+++ server.js
@@ -14,6 +14,8 @@ app.use(express.static("public"));
 // Routes
 require("./routes/apiRoutes")(app);
 require("./routes/htmlRoutes")(app);
+require("./routes/loginRoutes")(app);
+require("./routes/registerRoutes")(app);
 
 var syncOptions = { force: false };
